# hr_management/static/docs/swagger.yml
openapi: 3.0.0
info:
  title: HRMS Auth & User Service
  description: Comprehensive API documentation for the Human Resource Management System.
  version: "1.0.0"
servers:
  - url: /api/v1/auth
    description: Primary Auth Module Gateway
  - url: /api/v1/company-employee
    description: Company & Employee Module Gateway
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DesignationCreate:
      type: object
      required:
        - name
        - department
      properties:
        name:
          type: string
          example: Software Engineer
        department:
          type: integer
          example: 1
        status:
          type: string
          enum: [active, inactive]
          example: active
...
    EmployeeCreate:
      type: object
      required:
        - full_name
        - official_email_id
        - gender
      properties:
        full_name:
          type: string
          example: John Doe

        personal_email_id:
          type: string
          example: john@gmail.com

        official_email_id:
          type: string
          example: john@company.com

        password:
          type: string
          example: "123456"

        gender:
          type: string
          enum: [male, female, other]

        date_of_birth:
          type: string
          format: date

        branch:
          $ref: "#/components/schemas/Branch"

        department:
          $ref: "#/components/schemas/Department"

        designation:
          $ref: "#/components/schemas/Designation"

        shift:
          $ref: "#/components/schemas/Shift"

        attendence_policy:
          $ref: "#/components/schemas/AttendancePolicy"

        manager:
          type: integer
          example: 2

        phone_number:
          type: string
          example: "9876543210"

        status:
          type: string
          enum: [active, inactive]

        employee_documents:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeDocument"
...
    EmployeeDocument:
      type: object
      properties:
        document_type:
          type: integer
          example: 1
        document_file:
          type: string
          example: "base64/file/url"
    Branch:
      type: object
      properties:
        id:
          type: integer
          example: 1
    Department:
      type: object
      properties:
        id:
          type: integer
          example: 1
    Designation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    Shift:
      type: object
      properties:
        id:
          type: integer
          example: 1
    AttendancePolicy:
      type: object
      properties:
        id:
          type: integer
          example: 1
...
paths:
...
  /designation:
    get:
      summary: Get all designations
      description: Fetch list of designations with optional search, status filter, and pagination
      tags: [Designation]
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search by designation name
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
        - in: query
          name: paginate
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: List fetched successfully

    post:
      summary: Create designation
      tags: [Designation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesignationCreate"
      responses:
        "201":
          description: Designation created successfully

    delete:
      tags: [Designation]
      summary: Delete multiple designations (soft delete)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ids
              properties:
                ids:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
      responses:
        "200":
          description: Deleted successfully
        "400":
          description: No IDs provided or protected records
...
  /employee:
    get:
      summary: Get all employees
      tags: [Employee]
      description: Fetch employees with filters and pagination
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [active, inactive] }
        - in: query
          name: gender
          schema: { type: string, enum: [male, female, other] }
        - in: query
          name: department_id
          schema: { type: integer }
        - in: query
          name: designation_id
          schema: { type: integer }
        - in: query
          name: branch_id
          schema: { type: integer }
        - in: query
          name: employee_type_id
          schema: { type: integer }
        - in: query
          name: employee_status_id
          schema: { type: integer }
        - in: query
          name: paginate
          schema:
            type: string
            enum: ["true", "false"]

      responses:
        "200":
          description: Employee list fetched successfully

    post:
      summary: Create employee
      tags: [Employee]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeCreate"
      responses:
        "201":
          description: Employee created successfully

    delete:
      summary: Delete multiple employees (soft delete)
      tags: [Employee]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: integer
                  example: [1, 2]
      responses:
        "200":
          description: Deleted successfully
        "400":
          description: No IDs provided / Protected records
...